Extract tests for database migrations
Move existing tests to proper location.

ActiveUUID defines some methods to be used within database migrations,
and they need to be tested.
